+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-30 16:21:52
-// Commit: docs(i18n): update readme and API contracts for i18n
+// Updated at: 2026-05-30 16:34:16
+// Commit: test(i18n): implement widget integration tests for i18n
 
 class i18nComponentState {
-  final String commitRef = '1626';
-  final String lastMessage = 'docs(i18n): update readme and API contracts for i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-30T16:21:52+05:00');
+  final String commitRef = '1627';
+  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
+  final DateTime timestamp = DateTime.parse('2026-05-30T16:34:16+05:00');
 }